refactor: Feed 댓글 컴포넌트 세부적으로 모듈화 작업 및 기타 리팩토링
